% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_carbapenem_genes.R
\name{count_carb_gene_combinations_plot}
\alias{count_carb_gene_combinations_plot}
\title{Count Carbapenem Gene Combinations and Plot}
\usage{
count_carb_gene_combinations_plot(masterdata, output_dir, legend_size)
}
\arguments{
\item{masterdata}{A data frame containing the column \code{Bla_Carb_acquired}.}

\item{output_dir}{A character string specifying the directory to save the CSV and plot.}

\item{legend_size}{A number specifying the size of the legend.}
}
\value{
Invisibly returns a data frame with columns: Gene_Combination, Count, Percentage.
}
\description{
This function counts the occurrences of each unique combination of
carbapenem genes in the \code{Bla_Carb_acquired} column, keeping combinations intact,
calculates percentages, saves a CSV, and generates a bar plot.
The "-" entries are treated as "CARB-S".
}
\examples{
count_carb_gene_combinations_plot(masterdata, "~/git_repos/GenepiK/test_output/", legend_size= 20)
}
