% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ast_barplots.R
\name{create_ast_barplots}
\alias{create_ast_barplots}
\title{Create Barplots of Antimicrobial Susceptibility Testing (AST) by Isolate Type}
\usage{
create_ast_barplots(masterdata, output_dir, legend_size)
}
\arguments{
\item{masterdata}{A data frame containing the AST data. It must include the columns
'ghru_id', 'Isolate_type', and the specified antimicrobial columns.}

\item{output_dir}{A character string specifying the directory where the plot will be saved.}

\item{legend_size}{A number specifying the size of the legend.}
}
\value{
The function does not return a value. It prints a ggplot object to a PNG file
and provides a message confirming the save location.
}
\description{
This function generates a barplot to visualize the proportion of antimicrobial
susceptibility interpretations (Susceptible, Intermediate, Resistant)
for a set of drugs, faceted by 'Isolate_type'. The plot is saved as a PNG file.
}
\examples{
# Assuming 'masterdata' is a data frame loaded in the environment
# create_ast_barplots(masterdata = my_data, output_dir = "plots", legend_size= 20)
}
