% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ast_pivot.R
\name{create_ast_pivot}
\alias{create_ast_pivot}
\title{Create Pivot tables of Antimicrobial Susceptibility Testing (AST) by Isolate Type}
\usage{
create_ast_pivot(masterdata, output_dir)
}
\arguments{
\item{masterdata}{A data frame containing the AST data. It must include the columns
'ghru_id', 'Isolate_type', and the specified antimicrobial columns.}

\item{output_dir}{A character string specifying the directory where the plot will be saved.}
}
\value{
Invisibly returns a data frame with ST as rows and carbapenem genes as columns,
including row and column totals.
}
\description{
This function generates a pivot table to visualize the proportion of antimicrobial
susceptibility interpretations (Susceptible, Intermediate, Resistant)
for a set of drugs, by 'Isolate_type'. The tables are saved as a .CSV file.
}
\examples{
# Assuming 'masterdata' is a data frame loaded in the environment
pivot <- create_ast_pivot(masterdata, output_dir)
}
