% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ST_barplot.R
\name{create_ST_barplot}
\alias{create_ST_barplot}
\title{Create a Stacked Barplot of ST Distribution}
\usage{
create_ST_barplot(masterdata, output_dir, legend_size)
}
\arguments{
\item{masterdata}{A data frame containing at least \code{ghru_id}, \code{Isolate_type},
\code{ST}, and \code{Bla_Carb_acquired}.}

\item{output_dir}{A character string specifying the directory where the plot will be saved.}

\item{legend_size}{A numeric variable specifying the size of the axis text.}
}
\value{
The function does not return a value. It prints a ggplot object to a PNG file
and provides a message confirming the save location.
}
\description{
This function generates a stacked barplot showing the distribution of the
top 20 most frequent Sequence Types (STs) and their association with the
presence or absence of carbapenemase genes. The plot is saved as a PNG file.
}
\examples{
# Assuming 'masterdata' is a data frame loaded in the environment
# ST_barplot(masterdata = my_data, output_dir = "plots",legend_size= 20) 
}
