% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_data}
\alias{import_data}
\title{Check Columns and Read CSV Data}
\usage{
import_data(file_path, output_dir)
}
\arguments{
\item{file_path}{A character string specifying the path to the CSV file to be read.}

\item{output_dir}{A character string specifying the output directory. This parameter is currently used for a message and does not affect the function's output.}
}
\value{
The function does not return a value. Instead, it assigns a data frame to the \code{masterdata} variable
in the global environment and prints messages and warnings to the console.
}
\description{
This function reads a CSV file, checks if its column names match a predefined set of expected columns,
and then assigns the data frame to a global variable named \code{masterdata}. It provides messages
to the user about missing or unexpected columns.
}
\examples{
# Assuming 'data.csv' exists with the required columns
# and the appropriate packages (readr, dplyr, stringr) are loaded.
# import_data(file_path = "data.csv", output_dir = "my_output_folder")

}
