% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_st_counts.R
\name{topN_ST_counts_csv}
\alias{topN_ST_counts_csv}
\title{Create a table os top n ST Distribution}
\usage{
topN_ST_counts_csv(masterdata, output_dir, top_n)
}
\arguments{
\item{masterdata}{A data frame containing a column named \code{ST}.}

\item{output_dir}{A character string specifying the directory where the plot will be saved.}

\item{top_n}{An integer specifying how many top STs to include (default 10).}
}
\value{
Invisibly returns a data frame containing the top sequence types and percentages.
}
\description{
This function generates a stacked barplot showing the distribution of the
top 20 most frequent Sequence Types (STs) and their association with the
presence or absence of carbapenemase genes. The plot is saved as a PNG file.
}
\examples{
# Assuming 'masterdata' is a data frame loaded in the environment
topN_ST_counts_csv(masterdata, "~/git_repos/GenepiK/test_output/", 10)
}
